Correct 7 task
