+++ fast_api/src/models/film.py
@@ -1,20 +1,11 @@
 from __future__ import annotations
 
-import orjson
-
-# Используем pydantic для упрощения работы при перегонке данных из json в объекты
-from pydantic import BaseModel
-
 from models.genre import Genre
 from models.person import Person
+from models.parent_model import BaseOrjsonModel
 
 
-def orjson_dumps(v, *, default):
-    # orjson.dumps возвращает bytes, а pydantic требует unicode, поэтому декодируем
-    return orjson.dumps(v, default=default).decode()
-
-
-class Film(BaseModel):
+class Film(BaseOrjsonModel):
     id: str
     title: str
     description: str | None
@@ -23,8 +14,3 @@ class Film(BaseModel):
     actors: list[Person] | None
     writers: list[Person] | None
     directors: list[Person] | None
-
-    class Config:
-        # Заменяем стандартную работу с json на более быструю
-        json_loads = orjson.loads
-        json_dumps = orjson_dumps
