+++ fast_api/src/models/genre.py
@@ -1,19 +1,7 @@
 from __future__ import annotations
-
-import orjson
-from pydantic import BaseModel
+from models.parent_model import BaseOrjsonModel
 
 
-def orjson_dumps(v, *, default):
-    # orjson.dumps возвращает bytes, а pydantic требует unicode, поэтому декодируем
-    return orjson.dumps(v, default=default).decode()
-
-
-class Genre(BaseModel):
+class Genre(BaseOrjsonModel):
     id: str
     name: str | None
-
-    class Config:
-        # Заменяем стандартную работу с json на более быструю
-        json_loads = orjson.loads
-        json_dumps = orjson_dumps
