+++ fast_api/src/models/parent_model.py
@@ -1,0 +1,13 @@
+import orjson
+from pydantic import BaseModel
+
+
+def orjson_dumps(v, *, default):
+    # orjson.dumps возвращает bytes, а pydantic требует unicode, поэтому декодируем
+    return orjson.dumps(v, default=default).decode()
+
+
+class BaseOrjsonModel(BaseModel):
+    class Config:
+        json_loads = orjson.loads
+        json_dumps = orjson_dumps
