+++ fast_api/src/models/person.py
@@ -1,20 +1,11 @@
 from __future__ import annotations
 
-import orjson
-from pydantic import BaseModel
+from models.parent_model import BaseOrjsonModel
 
 
-def orjson_dumps(v, *, default):
-    return orjson.dumps(v, default=default).decode()
-
-
-class Person(BaseModel):
+class Person(BaseOrjsonModel):
     id: str
     full_name: str
-
-    class Config:
-        json_loads = orjson.loads
-        json_dumps = orjson_dumps
 
 
 class PersonDetailed(Person):
